Initialize all parameters for default message callback.
When libjpeg v9f is compiled with clang-17 using its memory sanitizer
(MSAN) it is possible to provoke a complaint when processing a specially
designed file like so:

 $ CC=clang-17 CFLAGS='-g -O0 -fno-omit-frame-pointer -fsanitize=memory' ./configure
 $ CC=clang-17 CFLAGS='-g -O0 -fno-omit-frame-pointer -fsanitize=memory' make -j2
 $ ./djpeg -outfile test 68404.jpg
   Independent JPEG Group's DJPEG, version 9f  14-Jan-2024
   Copyright (C) 2024, Thomas G. Lane, Guido Vollbeding
   ==686728==WARNING: MemorySanitizer: use-of-uninitialized-value
    #0 0x7fcfcd134a51 in format_message jerror.c:196:5
    #1 0x7fcfcd13390d in output_message jerror.c:104:3
    #2 0x7fcfcd1335e6 in emit_message jerror.c:139:7
    #3 0x7fcfcd0f2898 in next_marker jdmarker.c:1042:5
    #4 0x7fcfcd0f37c7 in read_markers jdmarker.c:1101:8
    #5 0x7fcfcd0e1a54 in consume_markers jdinput.c:565:11
    #6 0x7fcfcd0890ab in jpeg_consume_input jdapimin.c:318:15
    #7 0x7fcfcd088a3d in jpeg_read_header jdapimin.c:266:13
    #8 0x5619256288f6 in main djpeg.c:545:10
   [...]

While running this will execute next_marker() which passes two integer
parameters to WARNMS2(), which places the parameters in msg_parm.i[] and
calls the emit_message() callback. If jpeg_std_error() has been called
this points to the default format_message() callback, which will pass
all eight integer arguments in msg_parm to sprintf(). At this point MSAN
will complain because not all of msg_parm.i[] has been initialized, only
the first two integer parameters.

Similarly TRACEMS() called by get_soi() provokes a similar complaint if
djpeg is also passed the -verbose option. Since all the WARN*() and
TRACE*() macros finally end up in format_messages(), it is not surprising
that they may all provoke the same complaint.

This proposed patch avoids the MSAN complaint by clearing msg_parm.i[]
in jpeg_std_error(), so that later on, when format_message() is called,
all parameters in the array will have known values. I proposed a patch
accomplishing this by memset() to ijg, but Guido Vollbeding counter
proposed with a patch where each member is cleared separately. This will
be part of the upcoming libjpeg v10 release, but in the meantime we apply
this suggested upstream patch.